# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Task Manager API",
    "version": "1.0.0",
    "description": "REST API for the AI task manager — used by the web app, by agents through the MCP server, and by scripts.\n\n**Identity is attribution, not authentication.** Every request may send an `X-Actor` header (`agent:<name>` or `human:<name>`); it is recorded on every write and defaults to `human:anonymous`. Nothing verifies it — it exists so the timeline can tell an agent's change from a human's, and so agents can be kept from overwriting each other's claimed work.\n\n**Optional access gate.** A self-hosted server may set `API_TOKEN`; every `/api/v1` request must then carry `Authorization: Bearer <token>` (the `bearerAuth` scheme). Without it the API is open — right for localhost, wrong for anything reachable from a network you do not control.\n\n- Status changes go through `POST /api/v1/tasks/{taskId}/transition`, never PATCH. What a status requires is the shape of its payload.\n- Writes that accept `expectedVersion` fail with `VERSION_CONFLICT` when the task changed since it was read.\n- List responses are enveloped as `{ data, meta }`; single resources are returned directly. The events feed is cursor-paged instead.\n- Failures always use the `ErrorResponse` envelope with a `SCREAMING_SNAKE` `code`. Branch on the code, not on the status alone.\n- Dates are ISO 8601 UTC strings.\n- Task ids are also task numbers: task `42` renders as `TASK-000042`."
  },
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Only when the server sets API_TOKEN: every /api/v1 request must then send `Authorization: Bearer <API_TOKEN>`, or it gets 401 UNAUTHORIZED. /health and /docs stay open. It is one shared secret for every caller — who did what is still the X-Actor header."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "AT_LEAST_ONE_FIELD",
                  "UNAUTHORIZED",
                  "ACTOR_NOT_PERMITTED",
                  "TASK_NOT_FOUND",
                  "COMMENT_NOT_FOUND",
                  "VERSION_CONFLICT",
                  "TASK_ALREADY_CLAIMED",
                  "NOT_CLAIM_HOLDER",
                  "DEPENDENCY_CYCLE",
                  "NO_OPEN_DECISION",
                  "MALFORMED_JSON",
                  "PAYLOAD_TOO_LARGE",
                  "NOT_FOUND",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {},
              "requestId": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "The error envelope. Clients branch on error.code, never on the HTTP status alone."
      },
...
      "TaskFacets": {
        "type": "object",
        "properties": {
          "assignees": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "creators": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "assignees",
          "projects",
          "creators"
        ],
        "additionalProperties": false,
        "description": "Distinct non-null values actually present in the table. The assignee, project, and createdBy filters send values from here, which is what makes case-sensitive exact matching safe."
      },
      "TaskStats": {
        "type": "object",
        "properties": {
          "byStatus": {
            "type": "object",
            "properties": {
              "backlog": {
                "type": "integer",
                "minimum": 0
              },
              "needs_refinement": {
                "type": "integer",
                "minimum": 0
              },
              "todo": {
                "type": "integer",
                "minimum": 0
              },
              "in_progress": {
                "type": "integer",
                "minimum": 0
              },
              "blocked": {
                "type": "integer",
                "minimum": 0
              },
              "needs_user_decision": {
                "type": "integer",
                "minimum": 0
              },
              "needs_user_action": {
                "type": "integer",
                "minimum": 0
              },
              "needs_qa": {
                "type": "integer",
                "minimum": 0
              },
              "done": {
                "type": "integer",
                "minimum": 0
              },
              "deferred": {
                "type": "integer",
                "minimum": 0
              }
            }
          },
          "needsAttention": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "byStatus",
          "needsAttention"
        ],
        "additionalProperties": false,
        "description": "Task count per status — all ten keys present, zero included — plus needsAttention, the number of tasks waiting on a human (needs_user_decision + needs_user_action + needs_qa)."
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/tasks/facets": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Filter options present in the data",
        "description": "Declared before /tasks/{taskId} in the router, so `facets` is not captured as an id. The only safe source of values for the assignee, project, and createdBy filters, which match exactly.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
              "example": "agent:claude-code"
            },
            "required": false,
            "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
            "name": "X-Actor",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Distinct non-null assignees, projects, and creators, sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskFacets"
                }
              }
            }
          },
          "401": {
            "description": "The server runs with API_TOKEN set and the request carried no `Authorization: Bearer <token>`, or the wrong one. Never returned by a server without API_TOKEN.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Only the X-Actor header can fail here. `details` maps field name to messages. A malformed X-Actor header is also a VALIDATION_ERROR, with details[\"X-Actor\"].",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a requestId and never a stack trace.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Any other failure. Every error response in this API uses the ErrorResponse envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/stats": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Task counts per status",
        "description": "Every status is present in byStatus, zero included, so a client can index it without a fallback. needsAttention is the inbox size. Declared before /tasks/{taskId} for the same reason /facets is.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "pattern": "^[a-z0-9][a-z0-9._-]{0,63}$"
              },
              "minItems": 1,
              "description": "Repeatable. Count only tasks in these projects; omit to count every task."
            },
            "required": false,
            "description": "Repeatable. Count only tasks in these projects; omit to count every task.",
            "name": "project",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
              "example": "agent:claude-code"
            },
            "required": false,
            "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
            "name": "X-Actor",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Counts per status plus needsAttention.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStats"
                }
              }
            }
          },
          "401": {
            "description": "The server runs with API_TOKEN set and the request carried no `Authorization: Bearer <token>`, or the wrong one. Never returned by a server without API_TOKEN.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "An unknown parameter, or a project that is not a slug. `details` maps field name to messages. A malformed X-Actor header is also a VALIDATION_ERROR, with details[\"X-Actor\"].",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a requestId and never a stack trace.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Any other failure. Every error response in this API uses the ErrorResponse envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}